refactor lib folder in backend to contain a folder for logic and one for models
