+++ backend/lib/utilities/logic/getFileType_test.go
@@ -1,4 +1,4 @@
-package lib
+package logic
 
 import (
 	"backend/enums"
