+++ backend/lib/utilities/logic/getFilenameWIthoutExtension.go
@@ -1,4 +1,4 @@
-package lib
+package logic
 
 import (
 	"path"
