+++ backend/lib/utilities/logic/getFolderPath.go
@@ -1,4 +1,4 @@
-package lib
+package logic
 
 import (
 	"path/filepath"
