+++ backend/lib/utilities/logic/getFolderPath_test.go
@@ -1,4 +1,4 @@
-package lib
+package logic
 
 import (
 	"backend/models"
