+++ backend/lib/utilities/models/allowedFileTypes.go
@@ -1,4 +1,4 @@
-package lib
+package models
 
 import "slices"
